web/requirements.txt
flask>=3.0.0
requests>=2.31.0
yadisk>=3.4.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
